Scheduler: Fix R1 codebase
