+++ js/renovation/ui/scheduler/appointment_tooltip/utils/default_functions.tsx
@@ -1,6 +1,6 @@
 import { FormattedContent } from '../types';
 /* eslint-disable-next-line import/named */
-import { dxSchedulerAppointment } from '../../../../../ui/scheduler';
+import type { dxSchedulerAppointment } from '../../../../../ui/scheduler';
 
 export const defaultGetTextAndFormatDate = (): FormattedContent => ({ text: '', formatDate: '' });
 export const defaultGetSingleAppointment = (): dxSchedulerAppointment => ({});
