+++ js/renovation/ui/scheduler/model/utils.ts
@@ -1,0 +1,37 @@
+import { ViewType } from '../types';
+
+export const getAppointmentRenderingStrategyName = (viewType: ViewType): string => {
+  const appointmentRenderingStrategyMap: Record<ViewType, { renderingStrategy: string }> = {
+    day: {
+      renderingStrategy: 'vertical',
+    },
+    week: {
+      renderingStrategy: 'week',
+    },
+    workWeek: {
+      renderingStrategy: 'week',
+    },
+    month: {
+      renderingStrategy: 'horizontalMonth',
+    },
+    timelineDay: {
+      renderingStrategy: 'horizontal',
+    },
+    timelineWeek: {
+      renderingStrategy: 'horizontal',
+    },
+    timelineWorkWeek: {
+      renderingStrategy: 'horizontal',
+    },
+    timelineMonth: {
+      renderingStrategy: 'horizontalMonthLine',
+    },
+    agenda: {
+      renderingStrategy: 'agenda',
+    },
+  };
+
+  const { renderingStrategy } = appointmentRenderingStrategyMap[viewType];
+
+  return renderingStrategy;
+};
