+++ js/renovation/ui/scheduler/scheduler.tsx
@@ -10,7 +10,7 @@ import {
 } from '@devextreme-generator/declarations';
 import { TimeZoneCalculator } from './timeZoneCalculator/utils';
 // eslint-disable-next-line import/named
-import { Appointment } from '../../../ui/scheduler';
+import type { Appointment } from '../../../ui/scheduler';
 import { ViewProps, SchedulerProps } from './props';
 
 import { Widget } from '../common/widget';
@@ -25,7 +25,7 @@ import {
 } from './workspaces/types';
 import { WorkSpace } from './workspaces/base/work_space';
 import { SchedulerToolbar } from './header/header';
-import { getViewDataGeneratorByViewType } from '../../../ui/scheduler/workspaces/view_model/utils';
+import { getViewDataGeneratorByViewType } from '../../../__internal/scheduler/workspaces/view_model/m_utils';
 import type { AppointmentDataItem, DataAccessorType, DataSourcePromise } from './types';
 import {
   createDataAccessors, isViewDataProviderConfigValid,
