+++ js/renovation/ui/scheduler/timeZoneCalculator/createTimeZoneCalculator.ts
@@ -1,5 +1,5 @@
 import { TimeZoneCalculator } from './utils';
-import timeZoneUtils from '../../../../ui/scheduler/utils.timeZone';
+import timeZoneUtils from '../../../../__internal/scheduler/m_utils_time_zone';
 
 export const createTimeZoneCalculator = (
   currentTimeZone: string,
