+++ js/renovation/ui/scheduler/utils/data.ts
@@ -1,7 +1,7 @@
 import type { Appointment } from '../../../../ui/scheduler';
 import { AppointmentDataItem, DataAccessorType, LoadDataType } from '../types';
 import { replaceWrongEndDate } from '../../../../__internal/scheduler/appointments/data_provider/m_utils';
-import { createAppointmentAdapter } from '../../../../ui/scheduler/appointmentAdapter';
+import { createAppointmentAdapter } from '../../../../__internal/scheduler/m_appointment_adapter';
 import { TimeZoneCalculator } from '../timeZoneCalculator/utils';
 import { isDefined } from '../../../../core/utils/type';
