+++ js/renovation/ui/scheduler/view_model/to_test/views/utils/week.ts
@@ -6,7 +6,7 @@ import {
   getViewStartByOptions,
   setOptionHour,
 } from './base';
-import timeZoneUtils from '../../../../../../../ui/scheduler/utils.timeZone';
+import timeZoneUtils from '../../../../../../../__internal/scheduler/m_utils_time_zone';
 import { CalculateStartViewDate } from '../types';
 
 export const getIntervalDuration = (
@@ -15,7 +15,7 @@ export const getIntervalDuration = (
 
 export const getValidStartDate = (
   startDate: Date | undefined,
-  firstDayOfWeek: number,
+  firstDayOfWeek?: number,
 ): Date | undefined => (startDate
   ? dateUtils.getFirstWeekDate(startDate, firstDayOfWeek)
   : undefined);
